+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+- markdown: Add `WithListIndentStyle` option to control how blocks inside list items are indented.
+- cli: Add `-list-indent-style` flag to control this option from the CLI.
+
 ## v3.0.0 - 2022-12-14
 
 This is a new major release. You can install it using the module path: github.com/Kunde21/markdownfmt/v3.